Changed storing Nodes to indices of type uint16_t in opened nodes array, added using lamda as comparator to BinaryHeap.

+++ Spark/CUDA/PathFindingManager.h
@@ -47,7 +47,7 @@ namespace spark {
 		__host__ void findPathsCUDA() const;
 		__host__ void findPathsCPU() const;
 		__host__ std::uint16_t calculateNumberOfBlocks(std::uint16_t maxThreadsPerBlock) const;
-		__host__ std::uint8_t calculateNumberOfThreadsPerBlock(std::uint16_t numberOfBlocks) const;
+		__host__ std::uint16_t calculateNumberOfThreadsPerBlock(std::uint16_t numberOfBlocks) const;
 		__host__ std::deque<glm::ivec2> findPath(const glm::ivec2 startPoint, const glm::ivec2 endPoint) const;
 		__host__ void insertOrSwapNode(std::set<NodeAI>& openedNodes, const NodeAI& node) const;
 	};
